# Make bcolz_get_lcs_by_obj_ids even faster.