For retrieval of 1000 random light curves,

 1. Got rid of meta .where() statement, which dropped computation from 25s to 17s
 2. Got rid of Pandas things and used Numpy structured array methods directly, which made things take 9.4s
 3. Because of (1) we don't have to do chunking, which drops computation down to 4.7s